+++ examples/remoter/_request/exec.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-go run ./../../cmd/hls/hls-remoter/client/hls-remoter-cli -s localhost:8531 -f Bob
+go run ./../../cmd/hls/hls-remoter/client/hls-remoter-cli -s localhost:7531 -f Bob